tools/sprite-gen/requirements.txt
pillow>=10,<11
numpy>=1,<3
scipy>=1,<2
pyyaml>=6,<7
requests>=2,<3
py-aseprite>=0.0.2
responses>=0.25,<1
